+++ werobot/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.5.3'
+__version__ = '0.6.0'
 __author__ = 'whtsky'
 __license__ = 'MIT'